backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.20.0  # for tests

# Cache
redis>=5.0.0

# HTTP client
aiohttp>=3.9.0

# Data analysis
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0

# External data
yfinance>=0.2.36
fredapi>=0.5.0

# AI / LLM
anthropic>=0.40.0
google-genai>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Serialization
orjson>=3.10.0

# Logging
structlog>=24.0.0

# Utilities
websockets>=12.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
aioresponses>=0.7.6

# Linting
ruff>=0.4.0
mypy>=1.10.0
